Improve efficiency of flipping card game algorithm
Refactored the flipping card game algorithm in leetcode problem [822]. The previous implementation used set comprehension and itertools.chain, which made the algorithm more complex and less efficient. The refactored solution uses a simpler approach by storing duplicate numbers in a set and comparing potential answers with this set. If the number doesn't exist in the set and is smaller than the previous answer, it becomes the new answer. The new solution defaults the return value to