Add solution to the problem 'Queens That Can Attack the King' and '宝石补给'
This commit includes the solutions to the problems 'Queens That Can Attack the King' and '宝石补给'.
For 'Queens', an algorithm is implemented to find out the all the Queens which are in the attacking position of the King.
For '宝石补给', an algorithm is designed to perform a series of gifts and calculate the difference between the most and the least gemstones.

The commit also includes the calculations and rules used to find the solutions.

Related Topics: Array, Matrix, Simulation.