Add solutions for new LeetCode problems.
Added solutions for 3 new problems on LeetCode. The first problem is related to the movement of a robot represented on the number line, the second one involves rewarding the top K students, and the third one concerns stock price fluctuations. Each problem includes Python code to solve the problem, as well as a markdown file with the problem description. These solutions help users to better understand and solve similar problems.