Add python solution for finding single numbers in array
This update introduces a python solution to find the two single numbers in an array where all the other numbers occur twice. In this solution, we use a bit of operation and divide the numbers into two groups to achieve linearity and constant space complexity.

This solution is designed for 'leetcode' problem #260 'Single Number III'. It employs reduce() and xor from functools and operator modules respectively for implementing the algorithm and returns the two single numbers.

In addition, two new leetcode problem solutions are introduced: 'leetcode' problem #1726 'Tuple with Same Product' and 'leetcode' problem #2652 'Sum of Multiples'.